frontend/backend/requirements.txt
fastapi
requests
uvicorn
python-dotenv
Flask
google-generativeai
flask-cors 